CODA-74 - User-posts flow
