+++ controllers/api/user/get.go
@@ -4,16 +4,12 @@ import (
 	"github.com/coda-it/gowebapp/handlers"
 	"github.com/coda-it/gowebapp/models/user"
 	"github.com/coda-it/gowebapp/utils"
-	"github.com/coda-it/gowebserver/router"
 	"github.com/coda-it/gowebserver/session"
 	"github.com/coda-it/gowebserver/store"
 	"net/http"
 )
 
-// CtrUsers - api serving user data
-func CtrUsers(w http.ResponseWriter, r *http.Request, opt router.UrlOptions, sm session.ISessionManager, s store.IStore) {
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-
+func getHandler(w http.ResponseWriter, r *http.Request, sm session.ISessionManager, s store.IStore) {
 	links := map[string]map[string]string{
 		"self": map[string]string{
 			"href": "/api/user",
