+++ models/post/constants.go
@@ -1,0 +1,5 @@
+package post
+
+const (
+	CollectionName = "posts"
+)
