+++ models/user/constants.go
@@ -1,0 +1,5 @@
+package user
+
+const (
+	CollectionName = "users"
+)
